requirements.txt
streamlit
pdf2image
opencv-python
paddlepaddle
paddleocr
pandas
pillow